requirements.txt
# Core dependencies
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pandas==2.2.0
sqlalchemy==2.0.25
aiohttp==3.9.3
asyncio==3.4.3
tenacity==8.2.3
tqdm==4.66.1

# Clean architecture dependencies  
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Trading platform dependencies
PyQt5>=5.15.0
PyQtWebEngine>=5.15.0
plotly>=5.17.0
numpy>=1.24.0
pyqtgraph>=0.13.0
mplfinance>=0.12.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Web platform dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1